adding base project files.
Main.cpp is the main script in the current push which has all the core required. i have achieved a display rotation and transform of a triangle pyramid in the basic module. External libraries in use are GLEW-2.1.0 for running OpenGL commands efficiently, GLFW-3.3.8-WIN32 for simple interfaces, GLM which is a math library.